requirements.txt
configparser
psutil
pywin32; platform_system == "Windows"
